+++ teamKMS/src/main/java/com/devworker/kms/service/acl/AclService.java
@@ -1,9 +1,11 @@
 package com.devworker.kms.service.acl;
 
 import com.devworker.kms.dao.acl.AclDao;
+import com.devworker.kms.dic.PermissionType;
 import com.devworker.kms.dto.acl.AclDto;
 import com.devworker.kms.exception.NotExistException;
 import com.devworker.kms.repo.acl.AclRepo;
+import com.devworker.kms.util.AclUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -21,11 +23,13 @@ public class AclService {
 
 
     public String addAcl(AclDto dto) {
+        AclUtil.checkPermission(PermissionType.CREATEPERMISSION);
         aclRepo.save(dto.getDao());
         return dto.getAclId();
     }
 
     public void deleteAcl(String aclId) {
+        AclUtil.checkPermission(PermissionType.DELETEPERMISSION);
         aclRepo.findById(aclId).orElseThrow(() -> new NotExistException("Acl Not Exist"));
         aclRepo.deleteById(aclId);
     }
@@ -39,6 +43,7 @@ public class AclService {
     }
 
     public void updateAcl(AclDto dto) {
+        AclUtil.checkPermission(PermissionType.MODIFYPERMISSION);
         aclRepo.findById(dto.getAclId()).orElseThrow(() -> new NotExistException("Acl Not Exist"));
         aclRepo.save(dto.getDao());
     }
